+++ packages/components/pages/f-offers/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v2.0.0
+------------------------------
+*February 15, 2022*
+
+### Changed
+- Updated to include the new braze adapter
+- also updated to the latest version of content cards
+
+
 v1.15.2
 ------------------------------
 *February 13, 2022*